Stock_Prediction_Hybrid/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.7.0
tensorflow-cpu>=2.7.0
torch>=1.9.0
transformers>=4.11.0
yfinance>=0.1.63
textblob>=0.15.3
nltk>=3.6.3
plotly>=5.3.1
dash>=2.0.0
pandas-datareader>=0.10.0
beautifulsoup4>=4.9.3
requests>=2.26.0
python-dotenv>=0.19.0
streamlit>=1.2.0
PyYAML>=6.0.1 